#16 v3 (skończone)

+++ src/main/java/pl/put/poznan/transformer/buildings/LocINT.java
@@ -1,8 +1,8 @@
 package pl.put.poznan.transformer.buildings;
 
 public interface LocINT {
-    int     getArea();
-    int     getCapacity();
-    int     getLight();
-    float   getHeating();
+    double     getArea();
+    double     getCapacity();
+    double     getLight();
+    double   getHeating();
 }
